+++ lib/sbmt/kafka_consumer/base_consumer.rb
@@ -5,8 +5,9 @@ module Sbmt
     class BaseConsumer < Karafka::BaseConsumer
       class_attribute :skip_on_error, instance_writer: false, default: false
       class_attribute :middlewares, instance_writer: false, default: []
+      class_attribute :batch_middlewares, instance_writer: false, default: []
 
-      def self.consumer_klass(skip_on_error: nil, middlewares: nil)
+      def self.consumer_klass(skip_on_error: nil, middlewares: nil, batch_middlewares: nil)
         klass = Class.new(self) do
           def self.name
             superclass.name
@@ -16,6 +17,7 @@ module Sbmt
         # defaults are set in class_attribute definition
         klass.skip_on_error = skip_on_error if skip_on_error
         klass.middlewares = middlewares.map(&:constantize) if middlewares
+        klass.batch_middlewares = batch_middlewares.map(&:constantize) if batch_middlewares
         klass
       end
 
@@ -23,7 +25,7 @@ module Sbmt
         ::Rails.application.executor.wrap do
           if process_batch?
             with_batch_instrumentation(messages) do
-              process_batch(messages)
+              do_process_batch(messages)
               mark_message(messages.last)
             end
           else
@@ -99,6 +101,11 @@ module Sbmt
         end
       end
 
+      def do_process_batch(messages)
+        middleware_list = process_batch_middlewares + batch_middlewares
+        call_middlewares(messages, middleware_list) { process_batch(messages) }
+      end
+
       def do_consume(message)
         log_message(message) if log_payload?
 
@@ -107,7 +114,8 @@ module Sbmt
         message.payload
 
         with_common_instrumentation("process_message", message) do
-          call_middlewares(message, middlewares) { process_message(message) }
+          middleware_list = process_message_middlewares + middlewares
+          call_middlewares(message, middleware_list) { process_message(message) }
         end
 
         with_common_instrumentation("mark_as_consumed", message) do
@@ -147,7 +155,7 @@ module Sbmt
         message.payload || message.raw_payload
       end
 
-      def call_middlewares(message, middlewares)
+      def call_middlewares(payload, middlewares)
         return yield if middlewares.empty?
 
         chain = middlewares.map { |middleware_class| middleware_class.new }
@@ -156,7 +164,13 @@ module Sbmt
           if chain.empty?
             yield
           else
-            chain.shift.call(message, &traverse_chain)
+            middleware = chain.shift
+
+            if middleware.method(:call).arity == 2
+              middleware.call(payload, self, &traverse_chain)
+            else
+              middleware.call(payload, &traverse_chain)
+            end
           end
         end
         traverse_chain.call
@@ -172,6 +186,14 @@ module Sbmt
 
       def config
         @config ||= Sbmt::KafkaConsumer::Config.new
+      end
+
+      def process_message_middlewares
+        @process_message_middlewares ||= config.process_message_middlewares
+      end
+
+      def process_batch_middlewares
+        @process_batch_middlewares ||= config.process_batch_middlewares
       end
 
       def cooperative_sticky?