+++ spec/internal/app/middlewares/test_global_process_message_middleware.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class TestGlobalProcessMessageMiddleware
+  def call(_message)
+    yield
+  end
+end